Revert "Extract _patch_python_domain"
