+++ sphinx/ext/napoleon/__init__.py
@@ -7,7 +7,6 @@ from typing import TYPE_CHECKING
 
 import sphinx
 from sphinx.application import Sphinx
-from sphinx.ext.napoleon._patches import _patch_python_domain
 from sphinx.ext.napoleon.docstring import GoogleDocstring, NumpyDocstring
 from sphinx.util import inspect
 
@@ -340,6 +339,26 @@ def setup(app: Sphinx) -> ExtensionMetadata:
         'version': sphinx.__display_version__,
         'parallel_read_safe': True,
     }
+
+
+def _patch_python_domain() -> None:
+    from sphinx.domains.python._object import PyObject, PyTypedField
+    from sphinx.locale import _
+
+    for doc_field in PyObject.doc_field_types:
+        if doc_field.name == 'parameter':
+            doc_field.names = ('param', 'parameter', 'arg', 'argument')
+            break
+    PyObject.doc_field_types.append(
+        PyTypedField(
+            'keyword',
+            label=_('Keyword Arguments'),
+            names=('keyword', 'kwarg', 'kwparam'),
+            typerolename='class',
+            typenames=('paramtype', 'kwtype'),
+            can_collapse=True,
+        )
+    )
 
 
 def _process_docstring(
